v0.2.1: add tests for codes and hints

+++ src/main/java/dn/questenginev2/code/service/CodeServiceImpl.java
@@ -81,8 +81,12 @@ public class CodeServiceImpl implements CodeService {
     }
 
     @Override
-    public void deleteCode(Long codeId) {
+    public void deleteCode(Long codeId, Authentication auth) {
+        User currentUser = userService.getCurrentUser(auth);
+        questService.validateAuthorOrAdmin(currentUser);
         Code code = validateCodeExist(codeId);
+        questService.validateQuestAuthor(currentUser, code.getLevel().getQuest().getId());
+
         codeRepository.delete(code);
     }
